GenerateOutputFileTest class added

+++ RenewalLatterGenerator/Features/OutputFileHandler/GenerateOutputFile.cs
@@ -3,6 +3,9 @@
     using RenewalLatterGenerator.Common;
     using RenewalLatterGenerator.Models;
 
+    /// <summary>
+    /// Used to Generate invitation letter
+    /// </summary>
     public class GenerateOutputFile
     {
         public CustomerProduct CustomerProduct { get; set; }
@@ -11,9 +14,17 @@
 
         public string FilePath { get; set; }
 
+        /// <summary>
+        /// Start to generate the invitation letter
+        /// </summary>
         public void Start()
         {
-            FilePath = @"C:\Test\" + CustomerProduct.Id + "_" + CustomerProduct.FirstName + ".txt";
+            if (!FilePath.EndsWith("\\"))
+            {
+                FilePath = FilePath + "\\";
+            }
+
+            FilePath = FilePath + CustomerProduct.Id + "_" + CustomerProduct.FirstName + FileTypes.Text;
 
             var invitationTemplate = OutputTemplate.Get;
 
@@ -22,9 +33,18 @@
                 invitationTemplate = invitationTemplate.Replace(keyValue.Key, GetPropertyValue(CustomerProduct, keyValue.Value).ToString());
             }
 
-            FileSystem.WriteAllText(FilePath, invitationTemplate);
+            if (!FileSystem.FileExists(FilePath))
+            {
+                FileSystem.WriteAllText(FilePath, invitationTemplate);
+            }
         }
 
+        /// <summary>
+        /// Get property value by name
+        /// </summary>
+        /// <param name="source">source object</param>
+        /// <param name="propertyName">property name</param>
+        /// <returns>property value as object</returns>
         private static object GetPropertyValue(object source, string propertyName)
         {
             var propertyInfo = source.GetType().GetProperty(propertyName);
